Change pancake Recipe to protein version
Since i felt bad for not changing anything at all for the 2 new recipes,
i decided to remove the old vegan_pancakes2.html and rename it to
vegan_protein_pancakes. Also this way i had the chance to exercise
CLI commands which is more useful to me at this point.